Edit Cabin Func

+++ src/services/apiCabins.js
@@ -13,26 +13,38 @@ export async function getCabins() {
 
 //https://tidmrcrbhrqjcrtnnihm.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}-${Date.now()}`.replaceAll('/', '');
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const { data, error } = await supabase.from('cabins').insert([{ ...newCabin, image: imagePath }]);
+  let query = supabase.from('cabins');
+
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error('Cabin could not be created');
   }
 
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image);
 
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
-    console.error(storageError);
-    throw new Error('Cabin image could not be uploaded and the cabin was not created');
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id);
+      console.error(storageError);
+      throw new Error('Cabin image could not be uploaded and the cabin was not created');
+    }
   }
 
   return data;
